consolidated search functions
now there is a single searchHandler for both the zip and city search.  Based on which one is being activated, it fires the searchHandler function, which uses the name prop of the input being used to fire the appropriate function from the makeSearchTerm object.  If zip is being asked for, it makes the search term from makeSearchTerm.domestic[name=zip].  Plan is to add an 'allow international' button.