history bar colors lining up...a lot of trial and error with color-temperature plugin
color temperature plugin reads kelvin as a parameter.  had to find the right combo of coefficients on the original celcius data to get the colors to align with the temperatures in a sensible way.  blues don't look great in night mode, need to be more intense...reds maybe a little too intense in night mode.  day more looks good for both.  this is some duct tape slop if there ever was!  but its a small detail so its fine i guess